train is running

+++ CTCOffice/ctcoffice/CTCDispatch.h
@@ -26,7 +26,7 @@ class CTCDispatch
     public:
 
 
-        std::vector<float> authority;
+        std::vector<int> authority; // blocks so vector of ints
         float speed;
         std::vector<bool> tcStates;
         void setAuthority();
